Add GenerateAll alternative to GenerateFile
