+++ binapigen/run.go
@@ -16,76 +16,13 @@ package binapigen
 
 import (
 	"fmt"
-	"net/url"
 	"os"
 	"path"
 	"path/filepath"
 	"regexp"
 
 	"github.com/sirupsen/logrus"
-
-	"go.fd.io/govpp/binapigen/vppapi"
 )
-
-// VppInput defines VPP input parameters for the Generator.
-type VppInput struct {
-	ApiFiles   []*vppapi.File
-	VppVersion string
-}
-
-// ResolveVppInput resolves given input string into VppInput.
-//
-// Currently supported input formats are:
-// - directory with VPP API JSON files (e.g. /usr/share/vpp/api/)
-// - directory with VPP repository checkout (runs 'make json-api-files')
-func ResolveVppInput(input string) (*VppInput, error) {
-	vppInput := &VppInput{}
-
-	if input == "" {
-		input = vppapi.DefaultDir
-	}
-
-	u, err := url.Parse(input)
-	if err != nil {
-		logrus.Warnf("parsing url error: %v", err)
-	} else {
-		switch u.Scheme {
-		case "", "file":
-			info, err := os.Stat(input)
-			if err != nil {
-				return nil, fmt.Errorf("file error: %v", err)
-			} else {
-				if info.IsDir() {
-					apidir := vppapi.ResolveApiDir(u.Path)
-					logrus.Debugf("path %q resolved to api dir: %v", u.Path, apidir)
-
-					apiFiles, err := vppapi.ParseDir(apidir)
-					if err != nil {
-						logrus.Warnf("vppapi parsedir error: %v", err)
-					} else {
-						vppInput.ApiFiles = apiFiles
-						logrus.Infof("resolved %d apifiles", len(apiFiles))
-					}
-
-					vppInput.VppVersion = vppapi.ResolveVPPVersion(u.Path)
-					if vppInput.VppVersion == "" {
-						vppInput.VppVersion = "unknown"
-					}
-				} else {
-					return nil, fmt.Errorf("files not supported")
-				}
-			}
-		case "http", "https":
-			return nil, fmt.Errorf("http(s) not yet supported")
-		case "git", "ssh":
-			return nil, fmt.Errorf("ssh/git not yet supported")
-		default:
-			return nil, fmt.Errorf("unsupported scheme: %v", u.Scheme)
-		}
-	}
-
-	return vppInput, nil
-}
 
 func Run(vppInput *VppInput, opts Options, f func(*Generator) error) {
 	if err := run(vppInput, opts, f); err != nil {
@@ -137,6 +74,12 @@ func GeneratePlugins(genPlugins []string) func(*Generator) error {
 				if err := RunPlugin(p, gen, file); err != nil {
 					return err
 				}
+			}
+		}
+		for _, p := range genPlugins {
+			logrus.Debugf("  [GEN ALL plugin: %v]", p)
+			if err := RunPlugin(p, gen, nil); err != nil {
+				return err
 			}
 		}
 		return nil
